+++ src/frontend/components/feedbackBoardContainer.tsx
@@ -57,13 +57,14 @@ export interface FeedbackBoardContainerState {
   currentTeam: WebApiTeam;
   filteredProjectTeams: WebApiTeam[];
   filteredUserTeams: WebApiTeam[];
+  hasToggledArchive: boolean;
   isAppInitialized: boolean;
   isBackendServiceConnected: boolean;
   isReconnectingToBackendService: boolean;
   isSummaryDashboardVisible: boolean;
   isTeamDataLoaded: boolean;
   isAllTeamsLoaded: boolean;
-  maxvotesPerUser: number;
+  maxVotesPerUser: number;
   /**
    * Teams that the current user is specifically a member of.
    */
@@ -123,6 +124,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
       currentTeam: undefined,
       filteredProjectTeams: [],
       filteredUserTeams: [],
+      hasToggledArchive: false,
       isAllTeamsLoaded: false,
       isAppInitialized: false,
       isAutoResizeEnabled: true,
@@ -152,7 +154,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
       teamBoardDeletedDialogMessage: '',
       teamBoardDeletedDialogTitle: '',
       userTeams: [],
-      maxvotesPerUser: 5,
+      maxVotesPerUser: 5,
       feedbackItems: [],
       contributors: [],
       effectivenessMeasurementSummary: [],
@@ -162,7 +164,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
       allMembers: [],
       castedVoteCount: 0,
       boardColumns: [],
-      questionIdForDiscussAndActBoardUpdate: -1
+      questionIdForDiscussAndActBoardUpdate: -1,
     };
   }
 
@@ -219,7 +221,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
       });
 
       // Listen for signals for board updates.
-      reflectBackendService.onReceiveNewBoard(this.handleNewBoardAvailable);
+      reflectBackendService.onReceiveNewBoard(this.handleBoardCreated);
       reflectBackendService.onReceiveDeletedBoard(this.handleBoardDeleted);
       reflectBackendService.onReceiveUpdatedBoard(this.handleBoardUpdated);
     }
@@ -246,7 +248,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
 
   public componentWillUnmount() {
     window.removeEventListener('resize', this.handleResolutionChange);
-    reflectBackendService.removeOnReceiveNewBoard(this.handleNewBoardAvailable);
+    reflectBackendService.removeOnReceiveNewBoard(this.handleBoardCreated);
     reflectBackendService.removeOnReceiveDeletedBoard(this.handleBoardDeleted);
     reflectBackendService.removeOnReceiveUpdatedBoard(this.handleBoardUpdated);
   }
@@ -329,7 +331,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
     return formatter.format(value / 100);
   }
 
-  private readonly handleNewBoardAvailable = async (teamId: string, boardId: string) => {
+  private readonly handleBoardCreated = async (teamId: string, boardId: string) => {
     if (!teamId || this.state.currentTeam.id !== teamId) {
       return;
     }
@@ -423,8 +425,8 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
             currentBoard: null,
             isBoardUpdateDialogHidden: true,
             isTeamBoardDeletedInfoDialogHidden: false,
-            teamBoardDeletedDialogTitle: 'Retrospective deleted',
-            teamBoardDeletedDialogMessage: 'The retrospective you were viewing has been deleted by another user.',
+            teamBoardDeletedDialogTitle: 'Retrospective archived or deleted',
+            teamBoardDeletedDialogMessage: 'The retrospective you were viewing has been archived or deleted by another user.',
           }
         }
 
@@ -435,8 +437,8 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
           currentBoard: currentBoard,
           isBoardUpdateDialogHidden: true,
           isTeamBoardDeletedInfoDialogHidden: false,
-          teamBoardDeletedDialogTitle: 'Retrospective deleted',
-          teamBoardDeletedDialogMessage: 'The retrospective you were viewing has been deleted by another user. You will be switched to the last created retrospective for this team.',
+          teamBoardDeletedDialogTitle: 'Retrospective archived or deleted',
+          teamBoardDeletedDialogMessage: 'The retrospective you were viewing has been archived or deleted by another user. You will be switched to the last created retrospective for this team.',
         };
       }
 
@@ -558,7 +560,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
         ...recentVisitState,
         isTeamBoardDeletedInfoDialogHidden: false,
         teamBoardDeletedDialogTitle: 'Team not found',
-        teamBoardDeletedDialogMessage: 'Could not find the team specified in the url.',
+        teamBoardDeletedDialogMessage: 'Could not find the team specified in the URL.',
       }
 
       return {
@@ -608,7 +610,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
         ...queryParamTeamAndDefaultBoardState,
         isTeamBoardDeletedInfoDialogHidden: false,
         teamBoardDeletedDialogTitle: 'Board not found',
-        teamBoardDeletedDialogMessage: 'Could not find the board specified in the url.'
+        teamBoardDeletedDialogMessage: 'Could not find the board specified in the URL.'
       };
     }
   }
@@ -733,6 +735,21 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
     }
   }
 
+  private readonly handleArchiveToggle = (): void => {
+    this.setState({ hasToggledArchive: true });
+  };
+
+  // Handle when "Board" tab is clicked
+  private readonly handlePivotClick = async (item?: PivotItem): Promise<void> => {
+    if (item?.props.headerText === 'Board') { // Check if "Board" tab is clicked
+      if (this.state.hasToggledArchive) { // Reload only if archive was toggled
+        console.log('Reloading boards because archive state was toggled.');
+        await this.reloadBoardsForCurrentTeam();
+        this.setState({ hasToggledArchive: false }); // Reset the flag after reload
+      }
+    }
+  };
+
   /**
    * @description Loads all feedback boards for the current team. Defaults the selected board to
    * the most recently created board.
@@ -822,10 +839,10 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
     });
   }
 
-  private readonly createBoard = async (title: string, maxvotesPerUser: number, columns: IFeedbackColumn[], isIncludeTeamEffectivenessMeasurement: boolean, isBoardAnonymous: boolean, shouldShowFeedbackAfterCollect: boolean, displayPrimeDirective: boolean, permissions: IFeedbackBoardDocumentPermissions) => {
+  private readonly createBoard = async (title: string, maxVotesPerUser: number, columns: IFeedbackColumn[], isIncludeTeamEffectivenessMeasurement: boolean, isBoardAnonymous: boolean, shouldShowFeedbackAfterCollect: boolean, displayPrimeDirective: boolean, permissions: IFeedbackBoardDocumentPermissions) => {
     const createdBoard = await BoardDataService.createBoardForTeam(this.state.currentTeam.id,
       title,
-      maxvotesPerUser,
+      maxVotesPerUser,
       columns,
       isIncludeTeamEffectivenessMeasurement,
       isBoardAnonymous,
@@ -928,8 +945,8 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
     this.setState({ isRetroSummaryDialogHidden: true });
   }
 
-  private readonly updateBoardMetadata = async (title: string, maxvotesPerUser: number, columns: IFeedbackColumn[], isIncludeTeamEffectivenessMeasurement: boolean, isBoardAnonymous: boolean, shouldShowFeedbackAfterCollect: boolean, displayPrimeDirective: boolean, permissions: IFeedbackBoardDocumentPermissions) => {
-    const updatedBoard = await BoardDataService.updateBoardMetadata(this.state.currentTeam.id, this.state.currentBoard.id, maxvotesPerUser, title, columns, permissions);
+  private readonly updateBoardMetadata = async (title: string, maxVotesPerUser: number, columns: IFeedbackColumn[], isIncludeTeamEffectivenessMeasurement: boolean, isBoardAnonymous: boolean, shouldShowFeedbackAfterCollect: boolean, displayPrimeDirective: boolean, permissions: IFeedbackBoardDocumentPermissions) => {
+    const updatedBoard = await BoardDataService.updateBoardMetadata(this.state.currentTeam.id, this.state.currentBoard.id, maxVotesPerUser, title, columns, permissions);
 
     this.updateBoardAndBroadcast(updatedBoard);
   }
@@ -1082,7 +1099,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
           isDuplicatingBoard={isDuplicatingBoard}
           currentBoard={this.state.currentBoard}
           teamId={this.state.currentTeam.id}
-          maxvotesPerUser={this.state.maxvotesPerUser}
+          maxVotesPerUser={this.state.maxVotesPerUser}
           placeholderText={placeholderText}
           availablePermissionOptions={permissionOptions}
           onFormSubmit={onSubmit}
@@ -1354,7 +1371,7 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
           />
         </div>
         <div className="flex w-full items-center justify-start">
-          <Pivot>
+          <Pivot onLinkClick={this.handlePivotClick}>
             <PivotItem headerText="Board">
               {this.state.currentTeam && this.state.currentBoard && !this.state.isSummaryDashboardVisible &&
                 <div className="pivot-content-wrapper">
@@ -1696,11 +1713,11 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
                       className: 'retrospectives-dialog-modal',
                     }}>
                     <DialogContent>
-                      Are you sure you want to archive the retrospective `<i className="accent3">{this.state.currentBoard.title}</i>` and all of its feedback items?
+                      Are you sure you want to archive `<i>{this.state.currentBoard.title}</i>` retrospective board and its feedback?
                       <br /><br />
-                      <i className="accent3"><FontIcon iconName="LocationDot" /> Archived retrospectives are not visible in the board list.</i>
-                      <br /><br />
-                      <i className="accent3"><FontIcon iconName="LocationDot" /> No data will be deleted, Retrospective Board will be visible in the `History` tab.</i>
+                      <FontIcon iconName="LocationDot" /> <i>Archived retrospectives are removed from the board selection list.</i>
+                      <br />
+                      <FontIcon iconName="LocationDot" /> <i>However, boards and their feedback can be</i> restored <i>via the</i> `History` <i>tab.</i>
                     </DialogContent>
                     <DialogFooter>
                       <PrimaryButton onClick={this.archiveCurrentBoard} text="Archive" className="prime-directive-close-button" />
@@ -1712,7 +1729,11 @@ class FeedbackBoardContainer extends React.Component<FeedbackBoardContainerProps
             </PivotItem>
             <PivotItem headerText="History">
               <div className="pivot-content-wrapper">
-                <BoardSummaryTable teamId={this.state.currentTeam.id} supportedWorkItemTypes={this.state.allWorkItemTypes} />
+                <BoardSummaryTable
+                  teamId={this.state.currentTeam.id}
+                  supportedWorkItemTypes={this.state.allWorkItemTypes}
+                  onArchiveToggle={this.handleArchiveToggle}
+                />
               </div>
             </PivotItem>
           </Pivot>